refactor: modularize utils package
